change of level of projects with regard to pom.xml

+++ vehicles-registration/src/main/webapp/jsp/createInsurance.jsp
@@ -9,17 +9,23 @@
 		<th>&nbsp;</th>
 	</tr>
 	
+	<c:set var="i" value="1"/>
 	<c:forEach items="${createFormBean.insuranceList}" var="insurance">
 		<tr>
 			<fmt:parseDate value="${insurance.activeFrom}" pattern="dd-MM-yyyy" var="activeFrom"/>
 			<fmt:parseDate value="${insurance.activeTo}" pattern="dd-MM-yyyy" var="activeTo"/>
+			<fmt:formatDate value="${activeFrom}" pattern="d.M.yyyy" var="activeFrom"/>
+			<fmt:formatDate value="${activeTo}" pattern="d.M.yyyy" var="activeTo"/>
+			<spring:message code="confirm.insuranceDelete" arguments="${activeFrom},${activeTo}" var="confirmation"/>
 			
-			<td>${insurance.id}</td>
+			<td>${i}.</td>
 			<td>${insurance.companyName}</td>
-			<td class="center"><fmt:formatDate value="${activeFrom}" pattern="d.M.yyyy"/></td>
-			<td class="center"><fmt:formatDate value="${activeTo}" pattern="d.M.yyyy"/></td>
-			<td class="center"><a href="<%= request.getContextPath() %>/delete-insurance/${insurance.id}">x</a></td>
+			<td class="center">${activeFrom}</td>
+			<td class="center">${activeTo}</td>
+			<td class="center"><a href="<%= request.getContextPath() %>/delete-insurance/${insurance.id}" onclick="return confirm('${confirmation}')">x</a></td>
 		</tr>
+		
+		<c:set var="i" value="${i + 1}"/>
 	</c:forEach>
 </table>
